requirements.txt
newsapi-python
pdfkit
python-dotenv
tqdm
streamlit
streamlit-tags
pyppeteer
